3 changes made. New state machine for switch function, reversed direction of RotZ, Added code to allow 3DC default movements to be selected.
C001 - Reversed the default direction of RotZ
C002 - Added state machine to supress the default actions for two buttons when pressed together to cause a third action to occur. i.e. press the left and right button together to get the 3DC configuration menu to appear.
C003 - Added a global variable to switch between default 3DC movements as shown in the tutorials and the default movements provided by the Teaching Tech code.