fixed a bug introduced in the last release of code.
In the last release, I changed the operation of the key code and key report code to send a zero report at the end so that the 3DConnexion software knows when a button on the mouse has been released. While this worked I accidently introduced a bug where multiple keys could be sent in the same report when only the logical key was wanted. A simple one line fix was needed. This was in prior releases of the code but removed from the last one due to the change in operation. Now it is back